+++ .eslintrc
@@ -1,8 +1,12 @@
 {
   "extends": [
-    "@rocketseat/eslint-config/node"
+    "@rocketseat/eslint-config/node",
+    "plugin:vitest-globals/recommended"
   ],
   "rules": {
     "no-useless-constructor": "off"
+  },
+  "env": {
+    "vitest-globals/env": true
   }
 }
