+++ package.json
@@ -5,20 +5,23 @@
   "main": "index.js",
   "scripts": {
     "test": "vitest run",
-    "test:watch": "vitest"
+    "test:watch": "vitest",
+    "lint": "eslint src --ext .ts",
+    "lint:fix": "eslint src --ext .ts --fix"
   },
   "keywords": [],
   "author": "",
   "license": "ISC",
   "devDependencies": {
-    "@rocketseat/eslint-config": "^1.2.0",
-    "@types/node": "^20.3.2",
-    "eslint": "^8.43.0",
-    "typescript": "^5.1.3",
-    "vitest": "^0.32.2",
-    "vitest-tsconfig-paths": "^3.4.1"
+    "@rocketseat/eslint-config": "1.2.0",
+    "@types/node": "20.3.2",
+    "eslint": "8.43.0",
+    "eslint-plugin-vitest-globals": "1.3.1",
+    "typescript": "5.1.5",
+    "vitest": "0.32.2",
+    "vitest-tsconfig-paths": "3.4.1"
   },
   "dependencies": {
-    "dayjs": "^1.11.8"
+    "dayjs": "1.11.8"
   }
 }
