+++ src/domain/forum/application/repositories/answers-repository.ts
@@ -1,4 +1,4 @@
-import { Answer } from '../entities/answer'
+import { Answer } from '../../enterprise/entities/answer'
 
 export interface AnswersRepository {
   create(answer: Answer): Promise<void>
