+++ src/domain/forum/enterprise/entities/value-objects/slug.spec.ts
@@ -1,4 +1,4 @@
-import { expect, test } from 'vitest'
+import { test, expect } from 'vitest'
 import { Slug } from './slug'
 
 test('it should be able to create a new slug from text', () => {
